A big rewrite of controller logic, adding one more layer for raw values
* Added Raw, which is similar to Controller in terms of get/set/signalling,
  but stores raw u8 data. MIDI data goes into Raw unaltered. Raw -> Controller
  and Controller -> Raw converts data. UI talks to Controller (and in
  exceptional case also Raw, but this is device-specific logic) with get/set
  of converted data only;
* Rewrote effect_select logic using converted data "effect_select" and
  unconverted data "effect_select:raw";
* Added effect_tweak logic;

+++ core/src/model.rs
@@ -91,7 +91,7 @@ pub struct RangeControl { pub cc: u8, pub addr: u8, pub bytes: u8, pub from: u8,
     pub format: Format<Self> }
 #[derive(Clone, Debug)]
 pub struct Select { pub cc: u8, pub addr: u8,
-    pub from_midi: Option<Vec<u32>>, pub to_midi: Option<Vec<u32>> }
+    pub from_midi: Option<Vec<u16>>, pub to_midi: Option<Vec<u16>> }
